requirements.txt
anthropic>=0.40.0
claude-agent-sdk>=0.1.50
python-telegram-bot>=21.0
chromadb>=0.5.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0
